Corrections for zooming of trails. Solution for issue #18 and issue #21 - zoom_factor is corrected to its inverse (pixels per meter is inversely related to range) - don't use size_t in calculations but int.

+++ src/TrailBuffer.h
@@ -66,8 +66,8 @@ class TrailBuffer {
 
   RadarInfo *m_ri;
   size_t m_spokes;
-  size_t m_max_spoke_len;
-  size_t m_trail_size;
+  int m_max_spoke_len;
+  int m_trail_size;
   double m_previous_pixels_per_meter;
 
   TrailRevolutionsAge *m_true_trails;           // m_trails_size * m_trails_size
